# api.yaml
openapi: 3.0.3
info:
  title: Peerly API
  version: 1.0.0
paths:
...
  /api/weaves/generate:
    post:
      summary: Generate a new weave (AI-powered, Pro only)
      security: [{ clerk: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic]
              properties:
                topic:
                  type: string
                seed_nodes:
                  type: array
                  items:
                    type: string
                field:
                  type: string
                include_scaffolds:
                  type: boolean
                  default: true
      responses:
        "200":
          description: Created weave with nodes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Weave"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ProRequired"
  /api/weaves/{weaveId}/nodes:
    post:
      summary: Add a new node to a weave (Pro only)
      security: [{ clerk: [] }]
      parameters:
        - $ref: "#/components/parameters/weaveId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description]
              properties:
                title:
                  type: string
                description:
                  type: string
                contributed_by:
                  type: string
      responses:
        "200":
          description: Updated weave with gap detection result
          content:
            application/json:
              schema:
                type: object
                properties:
                  weave:
                    $ref: "#/components/schemas/Weave"
                  gap_detection:
                    $ref: "#/components/schemas/GapDetection"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ProRequired"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/weaves/{weaveId}/contribute:
    post:
      summary: Fill a scaffold node (Pro only)
      security: [{ clerk: [] }]
      parameters:
        - $ref: "#/components/parameters/weaveId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scaffold_node_id, title, description]
              properties:
                scaffold_node_id:
                  type: string
                title:
                  type: string
                description:
                  type: string
                contributed_by:
                  type: string
      responses:
        "200":
          description: Updated weave
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Weave"
        "400":
          description: Target is not a scaffold
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ProRequired"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Nodes ───────────────────────────────────────────────────────────────────
  /api/nodes/explain:
    post:
      summary: Generate an AI explanation for a node
      security: [{ clerk: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, topic]
              properties:
                title:
                  type: string
                description:
                  type: string
                topic:
                  type: string
                depth:
                  type: integer
                  default: 0
                difficulty:
                  type: integer
                  default: 1
                weaveId:
                  type: string
                nodeId:
                  type: string
      responses:
        "200":
          description: AI-generated explainer text
          content:
            application/json:
              schema:
                type: object
                properties:
                  explainer:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    clerk:
      type: http
      scheme: bearer
  parameters:
    weaveId:
      name: weaveId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unauthorized
...
    ProRequired:
      description: Pro plan required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: pro_required
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
  schemas:
    Node:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        depth:
          type: integer
        difficulty:
          type: integer
        is_scaffold:
          type: boolean
        contributed_by:
          type: string
          nullable: true
    Weave:
      type: object
      properties:
        id:
          type: string
        topic:
          type: string
        field:
          type: string
          nullable: true
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/Node"
    GapDetection:
      type: object
      properties:
        gap_detected:
          type: boolean
        missing_concept:
          type: string
          nullable: true
        scaffold_node:
          nullable: true
          $ref: "#/components/schemas/Node"
...